Consistent Sandbox termination (#137)
Always defer termination, and don't use the request context in the termination call.

Also consistently use a helper in tests.
