+++ modal-go/examples/sandbox-exec/main.go
@@ -27,7 +27,11 @@ func main() {
 		log.Fatalf("Failed to create Sandbox: %v", err)
 	}
 	log.Println("Started Sandbox:", sb.SandboxId)
-	defer sb.Terminate(ctx)
+	defer func() {
+		if err := sb.Terminate(context.Background()); err != nil {
+			log.Fatalf("Failed to terminate Sandbox %s: %v", sb.SandboxId, err)
+		}
+	}()
 
 	p, err := sb.Exec(ctx,
 		[]string{
