+++ modal-go/test/main_test.go
@@ -1,10 +1,12 @@
 package test
 
 import (
+	"context"
 	"os"
 	"testing"
 
 	modal "github.com/modal-labs/libmodal/modal-go"
+	"github.com/onsi/gomega"
 )
 
 // tc is the test Client, used for running tests against Modal infra.
@@ -19,4 +21,9 @@ func TestMain(m *testing.M) {
 
 	code := m.Run()
 	os.Exit(code)
+}
+
+func terminateSandbox(g *gomega.WithT, sb *modal.Sandbox) {
+	err := sb.Terminate(context.Background())
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
 }
